doc: added colorbars to plot_pwdslope
