+++ pylops/utils/_pwd2d_numba.py
@@ -1,13 +1,10 @@
-"""Numba-accelerated helpers for plane-wave destruction in 2D."""
-
-from __future__ import annotations
-
-__all__ = ["conv_allpass_numba"]
-
+from typing import Tuple
 
 from pylops.utils.typing import NDArray
 
-try:
+# Remap numba njit to no-ops and prange to range
+# if numba is not available
+try:  # pragma: no cover - executed only when numba is available
     from numba import njit, prange
 except ImportError:  # pragma: no cover - executed only without numba
 
@@ -22,7 +19,7 @@ except ImportError:  # pragma: no cover - executed only without numba
 
 
 @njit(fastmath=True, cache=True)
-def _B3(sigma: float) -> tuple[float, float, float]:
+def _B3(sigma: float) -> Tuple[float, float, float]:
     """Quadratic B-spline coefficients (3 taps)."""
     b0 = (1.0 - sigma) * (2.0 - sigma) / 12.0
     b1 = (2.0 + sigma) * (2.0 - sigma) / 6.0
@@ -31,7 +28,7 @@ def _B3(sigma: float) -> tuple[float, float, float]:
 
 
 @njit(fastmath=True, cache=True)
-def _B3d(sigma: float) -> tuple[float, float, float]:
+def _B3d(sigma: float) -> Tuple[float, float, float]:
     """Derivatives of quadratic B-spline coefficients."""
     b0 = -(2.0 - sigma) / 12.0 - (1.0 - sigma) / 12.0
     b1 = (2.0 - sigma) / 6.0 - (2.0 + sigma) / 6.0
@@ -40,7 +37,7 @@ def _B3d(sigma: float) -> tuple[float, float, float]:
 
 
 @njit(fastmath=True, cache=True)
-def _B5(sigma: float) -> tuple[float, float, float, float, float]:
+def _B5(sigma: float) -> Tuple[float, float, float, float, float]:
     """Quartic B-spline coefficients (5 taps)."""
     s = sigma
     b0 = (1 - s) * (2 - s) * (3 - s) * (4 - s) / 1680.0
@@ -52,7 +49,7 @@ def _B5(sigma: float) -> tuple[float, float, float, float, float]:
 
 
 @njit(fastmath=True, cache=True)
-def _B5d(sigma: float) -> tuple[float, float, float, float, float]:
+def _B5d(sigma: float) -> Tuple[float, float, float, float, float]:
     """Derivatives of quartic B-spline coefficients."""
     s = sigma
     b0 = (
@@ -94,10 +91,15 @@ def _B5d(sigma: float) -> tuple[float, float, float, float, float]:
 
 
 @njit(parallel=True, fastmath=True, cache=True)
-def conv_allpass_numba(
-    din: NDArray, dip: NDArray, order: int, u1: NDArray, u2: NDArray
+def _conv_allpass_numba(
+    din: NDArray,
+    dip: NDArray,
+    order: int,
+    u1: NDArray,
+    u2: NDArray,
 ) -> None:
-    """Numba kernel for PWD all-pass filtering used in :func:`pwd_slope_estimate`."""
+    """Numba kernel for PWD all-pass filtering used in
+    :func:`pylops.utils.signalprocessing.pwd_slope_estimate`."""
     n1, n2 = din.shape
     nw = 1 if order == 1 else 2
